+++ max_div/solver/_strategies/__init__.py
@@ -1,3 +1,2 @@
-from ._init_random import InitRandom
-from ._optim_dummy import OptimDummy
-from .core import ProgressTracker, SolverStrategy, StrategyType, hours, iterations, minutes, seconds
+from ._initialization import InitializationStrategy
+from ._optimization import OptimizationStrategy