+++ tests/solver/conftest.py
@@ -2,7 +2,9 @@ import numpy as np
 import pytest
 
 from max_div.solver import Constraint, DistanceMetric, DiversityMetric, MaxDivSolver, MaxDivSolverBuilder
-from max_div.solver._strategies import SolverStrategy
+from max_div.solver._duration import iterations, seconds
+from max_div.solver._solver_step import OptimizationStep
+from max_div.solver._strategies import InitializationStrategy, OptimizationStrategy
 
 
 @pytest.fixture
@@ -10,8 +12,11 @@ def example_problem_1() -> MaxDivSolver:
     # prepare data
     vectors = np.random.rand(10, 5).astype(np.float32)
     selection_size = 5
-    init_strategy = SolverStrategy.init_random()
-    optim_strategies = [SolverStrategy.optim_dummy(), SolverStrategy.optim_dummy()]
+    init_strategy = InitializationStrategy.random()
+    solver_steps = [
+        OptimizationStep(OptimizationStrategy.dummy(), seconds(1)),
+        OptimizationStep(OptimizationStrategy.dummy(), iterations(100)),
+    ]
     constraints = [
         Constraint(set(range(0, 5)), min_count=2, max_count=3),
         Constraint(set(range(5, 10)), min_count=2, max_count=3),
@@ -24,7 +29,7 @@ def example_problem_1() -> MaxDivSolver:
         builder.with_vectors(vectors)
         .with_selection_size(selection_size)
         .set_initialization_strategy(init_strategy)
-        .add_optimization_strategies(optim_strategies)
+        .add_solver_steps(solver_steps)
         .with_distance_metric(DistanceMetric.L1_MANHATTAN)
         .with_diversity_metric(DiversityMetric.min_separation())
         .with_constraints(constraints)